mock mapboxgl
