+++ .github/workflows/release.yml
@@ -132,5 +132,3 @@ jobs:
         run: npm run build
       - name: Publish to npmjs packages
         run: npm publish --access public
-        env:
-          NODE_AUTH_TOKEN: ${{ secrets.NPM_TOKEN }}
