+++ .github/workflows/test.yml
@@ -88,5 +88,3 @@ jobs:
         run: npm ci
       - name: Test
         run: npm run test
-        env:
-          STORYBOOK_MAPBOX_TOKEN: ${{ secrets.STORYBOOK_MAPBOX_TOKEN }}
